--- server/db/dashboard.ts
@@ -1,20 +1,0 @@
-import { connection } from "./mysql-conn";
-
-export class dashboard {
-
-    public async getGrades(){
-        try{
-            return await connection.query(
-                'select * from grade',
-                function(err: any, results: any, fields: any) {
-                  console.log(results);
-                  console.log(fields);
-                }
-              );
-        }
-        catch(err: any){
-
-        }
-    }
-
-}
