--- server/db/mysql-conn.ts
@@ -1,9 +1,0 @@
-
-const mysql = require('mysql2');
-
-export const connection = mysql.createConnection({
-  host: 'https://remotemysql.com/',
-  user: 'svQtxIxilZ',
-  database: 'svQtxIxilZ',
-  password : 'Z5uzX9DkGm',
-});
